requirements.txt
Django>=4.2.0
djangorestframework
django-cors-headers
Pillow
